tried to load a symbol for each selector from the symboltable. this seems to work quite well, but somehow the comparison between the message and selector in invoke still fails, even if they look the same as "%p".

+++ src/kernel/collection/bucket.h
@@ -7,7 +7,7 @@
 
 struct Bucket {
     VariableHeader      header;
-    SmallInteger        tally;
+    int                 tally;
     Object              value[];
 };
